# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para plataforma de serviços de limpeza.",
        "title": "Phresh-Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/agenda/bloqueios": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agenda"
                ],
                "summary": "Listar bloqueios do faxineiro autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RespostaBloqueio"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agenda"
                ],
                "summary": "Criar bloqueio pessoal de agenda",
                "parameters": [
                    {
                        "description": "Dados do bloqueio",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoBloqueio"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaBloqueio"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/agenda/bloqueios/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agenda"
                ],
                "summary": "Remover bloqueio pessoal de agenda",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do bloqueio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/agenda/disponibilidades": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agenda"
                ],
                "summary": "Listar disponibilidades do faxineiro autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RespostaDisponibilidade"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agenda"
                ],
                "summary": "Adicionar bloco de disponibilidade",
                "parameters": [
                    {
                        "description": "Dados da disponibilidade",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoDisponibilidade"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaDisponibilidade"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/agenda/disponibilidades/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agenda"
                ],
                "summary": "Remover bloco de disponibilidade",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da disponibilidade",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/avaliacoes": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "avaliacoes"
                ],
                "summary": "Criar avaliação de um serviço concluído",
                "parameters": [
                    {
                        "description": "Dados da avaliação",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoCriarAvaliacao"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaAvaliacao"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/faxineiros/{faxineiro_id}/avaliacoes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "avaliacoes"
                ],
                "summary": "Listar avaliações de um faxineiro",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do faxineiro",
                        "name": "faxineiro_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RespostaAvaliacao"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/faxineiros/{faxineiro_id}/estatisticas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "avaliacoes"
                ],
                "summary": "Buscar estatísticas de avaliações de um faxineiro",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do faxineiro",
                        "name": "faxineiro_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaEstatisticasFaxineiro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.RequisicaoBloqueio": {
            "type": "object",
            "properties": {
                "data_fim": {
                    "type": "string"
                },
                "data_inicio": {
                    "type": "string"
                }
            }
        },
        "dto.RequisicaoCriarAvaliacao": {
            "type": "object",
            "properties": {
                "comentario": {
                    "type": "string"
                },
                "limpeza_id": {
                    "type": "integer"
                },
                "nota": {
                    "type": "integer"
                }
            }
        },
...
        "dto.RequisicaoDisponibilidade": {
            "type": "object",
            "properties": {
                "dia_semana": {
                    "description": "0=Domingo ... 6=Sábado",
                    "type": "integer"
                },
                "hora_fim": {
                    "description": "1-24",
                    "type": "integer"
                },
                "hora_inicio": {
                    "description": "0-23",
                    "type": "integer"
                }
            }
        },
...
        "dto.RespostaAvaliacao": {
            "type": "object",
            "properties": {
                "cliente_id": {
                    "type": "integer"
                },
                "comentario": {
                    "type": "string"
                },
                "faxineiro_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "limpeza_id": {
                    "type": "integer"
                },
                "nota": {
                    "type": "integer"
                }
            }
        },
        "dto.RespostaBloqueio": {
            "type": "object",
            "properties": {
                "data_fim": {
                    "type": "string"
                },
                "data_inicio": {
                    "type": "string"
                },
                "e_pessoal": {
                    "type": "boolean"
                },
                "faxineiro_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "solicitacao_id": {
                    "type": "integer"
                }
            }
        },
        "dto.RespostaDisponibilidade": {
            "type": "object",
            "properties": {
                "dia_semana": {
                    "type": "integer"
                },
                "faxineiro_id": {
                    "type": "integer"
                },
                "hora_fim": {
                    "type": "integer"
                },
                "hora_inicio": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "dto.RespostaErro": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "detalhes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mensagem": {
                    "type": "string"
                }
            }
        },
        "dto.RespostaEstatisticasFaxineiro": {
            "type": "object",
            "properties": {
                "faxineiro_id": {
                    "type": "integer"
                },
                "media_nota": {
                    "type": "number"
                },
                "total_avaliacoes": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    }
}